+++ Gestion_Usuarios/src/main/java/com/gestionusuario/app/enumerator/PermisosLista.java
@@ -2,9 +2,9 @@ package com.gestionusuario.app.enumerator;
 
 public class PermisosLista  {
 
-	public static final int CreadorPerfil = 1;
-	public static final int ModificadorPerfil = 2;
-	public static final int DesactivadorPerfil = 3;
-	public static final int AsignadorPerfil = 4;
+	public static final Long CreadorPerfil = (long) 1;
+	public static final Long ModificadorPerfil = (long) 2;
+	public static final Long DesactivadorPerfil = (long) 3;
+	public static final Long AsignadorPerfil = (long) 4;
 	
 }
