+++ Gestion_Usuarios/src/main/java/com/gestionusuario/app/service/impl/PerfilServiceImpl.java
@@ -53,7 +53,7 @@ public class PerfilServiceImpl implements PerfilService{
 	}
 
 	@Override
-	public int ValidarPermisos(Integer idUsuario, int idPermiso) throws Exception {
+	public int ValidarPermisos(Long idUsuario, Long idPermiso) throws Exception {
 	
 		try {
 			return this.perfildao.ValidarPermisos(idUsuario, idPermiso);
